[opentitantool] Factored UsbBackend into separate util file
The logic for finding and interacting with a USB device was somewhat
common between the CW310 and HyoerDebug implementation.  Furthermore,
separating the low-level USB parts into a struct, allows us to later
make it a trait, which could have separate mocked implementations,
allowing for unit testing of the various backends.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I6b6af1b53ee9512479cc0089381d68e9d21d3739