[opentitantool] Support configuration of GPIO pin modes
Replace GPIO direction enum (in/out) with a mode enum
(input/push pull/open drain).

Also, introduce functionality in TransportWrapper to iterate over pin
declarations and apply mode and default output level on a set of pins.
This is intended to be used both on a broad set of pins, as well as
multiple named combinations on a smaller set of pins strappings.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ia6b9f392be2c34451bcdfadf82b8864797d5553b