[otbn,dv] Teach otbn_model to pass validity bits to and from the ISS
The binary file format has changed and now consists of 5 bytes for
each 32-bit word. The first byte is a "validity byte", which should
have value 0 (for invalid) or 1 (for valid). The other 4 bytes are a
little-endian encoded 32-bit word.

The first change in otbn_model is to use Ecc32MemArea's new
ReadWithIntegrity method to get the integrity bits as well as the raw
data. We then pass this to the ISS in the new format. In the other
direction (reading data from the ISS), we have to parse the new format
and also need to tweak our correctness check.

Note that there's a temporary hack (with TODO message) here, which
will go away once we've implemented integrity bits properly on the ISS
side.

On the ISS side, the only work is to read and write the new file
format (with struct format string "<BI") and fix up the various bits
of packing logic. There's a corresponding hack (also with TODO
message), reminding us to handle validity bits properly when loading
DMEM.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>