[dv/alert_handler] fix alert ping timeout seq regression error
This PR fixes the timeout issue in nightly regression regarding
alert_handler_ping_timeout_vseq.
The issue is that we locked the ping_en register via regwen before
enabling the ping_en csr.
The solution is to move the order of regwen write task.

This PR also sets the min number of alert_en to be NUM_ALERTS-4 to avoid
this sequence running too long.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>