[spi_device] Add write mask to RxFifo
In the issue #9119, I decide to revise the design to support BitMask in
the Generic mode.

When Generic mode (a.k.a FwMode) was implemented, DPSRAM supported
full-word only. After introducing the Flash / Passthrough mode, the
DPSRAM is replaced to Bit-wise mask enabled DPSRAM.

As discussed in #9119, the RXF does "Read-Modify-Write" to support
partial write. With bit-wise mask, the RXF FSM becomes much simpler by
eliminating RMW operations.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>