[sw/ottf] Optimize OTTF for non-concurrent tests.
Concurrency tests require a thin OS layer (e.g., FreeRTOS) to be able to
spawn multiple tasks to stress the hardware. However, the majority of
chip-level tests are sequential, and do not require any concurrency support
from the OTTF. To minimize simulation runtimes for these tests, a
configuration option is added to the OTTF test configuration struct to allow
test developers to run their tests: 1) on bare-metal or 2) as a FreeRTOS task.

This partially addresses #8015, which proposes refactoring the on-device test
framework to support concurrency testing.

Signed-off-by: Timothy Trippel <ttrippel@google.com>