[util] Move some generic parsing code to util
This is used in the OTBN code to convert dictionaries parsed from YAML
or hjson into proper objects. This patch moves the code to a new
util/serialize directory and teaches the OTBN code that was using it
how to find the new version.

The idea is that this might be useful for tightening up parsing in
e.g. reggen and topgen and will hopefully also be useful for future
tooling.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>