[dv/prim_alert_receiver] Fix assertion that consumes large mem
This PR fixes an assertion sequence with large unbounded delays that
consumes large memory runtime.
If I understand correctly, we do not need to wait every clock cycle in
the init state for `init_trig_i` to change to True. We only need the
condition that we are in the init state and `init_trig_i` changes from
true to false, and in which we have an alert signal integrity error.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>