[sw/ottf] Fix processing of `mepc` on synchronous IRQ.
When an IRQ is fired on RISC-V processors, the PC is stored in the `mepc`
register. This address is then used as the return address when the `mret`
instruction is called upon returning from an ISR.

When the IRQ is synchronous, e.g. an exception, the `mepc` address must be
updated to point to the **_next_** instruction **_after_** the instruction
that caused the exception, otherwise we will get stuck in a continuous
exception cycle.

Since we support RISC-V compressed instructions, we must first check whether
the faulting instruction is compressed (i.e., 16-bits) or not (i.e., 32-bits)
in order to know what offset (2 or 4) to add to the `mepc` before returning
from the ISR.

Until now, the OTTF ISR handlers assume the trapped instruction was
uncompressed. This fixes #9208.

Signed-off-by: Timothy Trippel <ttrippel@google.com>