[ spi_host, rtl ] Properly handle back-to-back segments
Issue #9285 uncovers a bug whereby back-to-back segments are not being properly handled in the rtl.

The issue is that when processing back to back segments, the FSM needs to carefully choose which copy
of the segment parameters to look at.  For anything initializing states for the following segment the
FSM needs to look at the new incoming segment parameters.  For finalizing processing of the previous segment
(e.g., flushing data out of the shift register) the FSM needs to look at the previous parameters.

This commit manages this subtlety.

Fixes #9285 (with patch to test bench in @rasmus-madsen's offline TB, see issue for patch)

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>