[aes, pre_sca] Enable formal verification of more than a single S-Box
This commit enables the formal verification of either the full SubBytes,
i.e., the 16 S-Boxes in the data path, and even a reduced AES round
containing also ShiftRows and MixColumns. This is useful as formal
verification of the entire AES cipher core is currently out of scope
from a tool run time point of view whereas SCA leakage outside the
reduced AES round is unlikely.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>