[reggen] Headers for non-homogenous multiregs
Rather than emitting a warning we can generate the defines for
non-homogeneous multiregs, with the exception of the NUMBER_OF_*_PER_REG
field which doesn't appear to be functioning properly or to be used by
OpenTitan.

(A similar structure is used in pmp.c where it's defined without any
autogeneration, These mostly indicate that there are 32 of a certian
kind of 1 bit field in a way that doesn't appear to get initialized or
really to make sense, so I'm fairly certian it's done in error.)

Removed apparently unused and poorly behaved NUMBER_OF_*_PER_REG from
the headers until we have a usecase to add and validate it.

Signed-off-by: Drew Macrae <drewmacrae@google.com>