[top] Temporarily bypass ast main_pok in fpga
Ast's main_pok does not appear to be working in fpga mode.
I have not investigated closely but it appears to be related
to incorrect ifdef's used during fpga build, which results in
the signal being a permanent '1'.

This commit mainly is there to enable low power transitions for
FPGA and thus will work around it for now.

Signed-off-by: Timothy Chen <timothytim@google.com>