[mask_rom] Add OTP option to enable AST and reduce AST_CFG size
Using this patch AST initialization will be skipped unless the new
OTP value AST_INIT_EN is set to true (0xA). This new value can be
used as a 'tag' to show that the AST configuration in OTP is valid.
We may eventually want to turn this into some sort of checksum.

While we are here reduce the OTP AST_CFG field size to 32 words to
match the configuration.

Updates #9296.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>